importa ListItem e ListItemText para primeira composição da lista
